refactor: helplessly try to make the gui conform to the 3domain architecture more
